Change SD access to use more generic functions
