+++ hydrabus/hydrabus_sd.h
@@ -1,0 +1,23 @@
+/*
+ * HydraBus/HydraNFC
+ *
+ * Copyright (C) 2014-2017 Benjamin VERNOUX
+ * Copyright (C) 2017 Nicolas OBERLI
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "common.h"
+
+int cmd_sd(t_hydra_console *con, t_tokenline_parsed *p);
+int cmd_show_sd(t_hydra_console *con);
